refactor: 컨트롤러 분리 및 DateRange 유틸 분리, API 응답 형식 개선
