+++ src/controllers/HabitWriteController.js
@@ -1,78 +1,5 @@
 import prisma from '../lib/prisma.js';
-
-// 습관 조회
-export const getTodayHabits = async (req, res) => {
-  try {
-    const { studyId } = req.params;
-
-    const study = await prisma.study.findUnique({
-      where: {
-        id: Number(studyId),
-      },
-      select: {
-        title: true,
-      },
-    });
-
-    if (!study) {
-      return res.status(404).json({
-        success: false,
-        message: '스터디를 찾을 수 없습니다.',
-        errors: [],
-      });
-    }
-    // today 추후 toggleHabit 파트랑 뭉쳐서 빼버리기
-    const todayStart = new Date();
-    todayStart.setHours(0, 0, 0, 0);
-
-    const tomorrowStart = new Date(todayStart);
-    tomorrowStart.setDate(tomorrowStart.getDate() + 1);
-
-    const habits = await prisma.habit.findMany({
-      where: {
-        studyId: Number(studyId),
-      },
-      include: {
-        records: {
-          where: {
-            date: {
-              gte: todayStart,
-              lt: tomorrowStart,
-            },
-          },
-          select: {
-            isCompleted: true,
-          },
-        },
-      },
-      orderBy: {
-        id: 'asc',
-      },
-    });
-
-    const habitIds = habits.map((h) => ({
-      id: h.id,
-      name: h.name,
-      isCompleted: h.records[0]?.isCompleted ?? false,
-    }));
-
-    return res.status(200).json({
-      success: true,
-      message: '오늘의 습관 목록 조회에 성공했습니다.',
-      data: {
-        studyTitle: study.title,
-        habits: habitIds,
-      },
-    });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({
-      success: false,
-      message: '서버 내부 오류가 발생했습니다.',
-      errors: [],
-    });
-  }
-};
+import { getTodayRange } from '../lib/DateRange.js';
 
 // 습관 생성
 export const createHabit = async (req, res) => {
@@ -150,6 +77,8 @@ export const toggleHabit = async (req, res) => {
   try {
     const { studyId, habitId } = req.params;
 
+    const { start: todayStart, end: tomorrowStart } = getTodayRange();
+
     const habit = await prisma.habit.findFirst({
       where: {
         id: Number(habitId),
@@ -165,13 +94,6 @@ export const toggleHabit = async (req, res) => {
       });
     }
 
-    // 추후 조회 API 내 today ~ 와 함께 빼기
-    const todayStart = new Date();
-    todayStart.setHours(0, 0, 0, 0);
-
-    const tomorrowStart = new Date(todayStart);
-    tomorrowStart.setDate(tomorrowStart.getDate() + 1);
-
     const record = await prisma.habitRecord.findFirst({
       where: {
         habitId: Number(habitId),
@@ -182,15 +104,17 @@ export const toggleHabit = async (req, res) => {
       },
     });
 
+    let updatedRecord;
+
     if (record) {
-      await prisma.habitRecord.update({
+      updatedRecord = await prisma.habitRecord.update({
         where: { id: record.id },
         data: {
           isCompleted: !record.isCompleted,
         },
       });
     } else {
-      await prisma.habitRecord.create({
+      updatedRecord = await prisma.habitRecord.create({
         data: {
           habitId: Number(habitId),
           date: todayStart,
@@ -203,8 +127,8 @@ export const toggleHabit = async (req, res) => {
       success: true,
       message: '습관 완료 상태가 변경되었습니다.',
       data: {
-        habitId,
-        isCompleted: true,
+        habitId: Number(habitId),
+        isCompleted: updatedRecord.isCompleted,
       },
     });
   } catch (error) {
