+++ src/lib/DateRange.js
@@ -1,0 +1,9 @@
+export const getTodayRange = () => {
+  const todayStart = new Date();
+  todayStart.setHours(0, 0, 0, 0);
+
+  const tomorrowStart = new Date(todayStart);
+  tomorrowStart.setDate(tomorrowStart.getDate() + 1);
+
+  return { start: todayStart, end: tomorrowStart };
+};
